requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
logging==0.4.9.6
requests==2.31.0
humanize==4.8.0
psutil==5.9.8
